requirements.txt
tweepy==4.14.0
   flask==2.3.3
   python-dotenv==1.0.0
   flask-cors==4.0.0
   gunicorn==21.2.0